chore(test-utils): refactor API mocks to more digestible chunks
* Split the API mocks by topic (stats, auth, etc.)
